Add aiori-DAOS
This patch adds an initial aiori-DAOS implementation based on DAOS
POSIX.  Not all IOR options are applicable to DAOS, but this should
work:

  mpirun -np 8 ior -a DAOS -w -W -r -R -t 2m -b 32m -y 4

In this case, a container will be created with 8 shards (round-robin
over all targets) and used for a write, a writeCheck, a read, and a
readCheck test.  In all tests, for each 2 MB "xfer", each process will
submit 4 asynchronous I/Os and wait for the completions before starting
next xfer.  All write I/Os form a single epoch, committed at container
close time.

Signed-off-by: Li Wei <wei.g.li@intel.com>