Add "daosobjects" option
This patch adds a new "-O" option, "daosobjects", which specifies how
many DAOS objects should be involved.  Currently, the number of
processes in a run must be a multiple of or equal to the number of
objects.  That is, either each process accesses its own object or the
same number of processes access each object in a block-interleaved way.
In addition, segments are now supported.

Signed-off-by: Li Wei <wei.g.li@intel.com>